+++ libs/taskbox/src/app/app.component.ts
@@ -4,7 +4,7 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-root',
   template: `
-   <app-inbox-screen></app-inbox-screen>
+   <lda-inbox-screen></lda-inbox-screen>
  `,
 })
 export class AppComponent {
