+++ libs/taskbox/src/app/app.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
-import {TaskModule} from "../component/task.module";
+import {TaskModule} from "../task/task.module";
 import {NgxsModule} from "@ngxs/store";
 import {environment} from "../environments/environment";
 import {NgxsReduxDevtoolsPluginModule} from "@ngxs/devtools-plugin";
 import {NgxsLoggerPluginModule} from "@ngxs/logger-plugin";
-import InboxScreenComponent from "../component/inbox-screen.component";
-import PureInboxScreenComponent from "../component/pure-inbox-screen.component";
+import InboxScreenComponent from "../task/inbox-screen.component";
+import PureInboxScreenComponent from "../task/pure-inbox-screen.component";
 
 @NgModule({
   declarations: [
