+++ libs/taskbox/src/button/button.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+
+import ButtonComponent from "./button.component";
+
+@NgModule({
+  imports: [CommonModule, ButtonComponent],
+  exports: [ButtonComponent],
+  declarations: [],
+})
+export class ButtonModule {}
