+++ libs/taskbox/src/component/components.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ButtonModule } from '../button';
+import {ButtonModule} from "../button/button.module";
 import PageComponent from "../pages/page.component";
 import HeaderComponent from "../header/header.component";
